requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
sqlalchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
psycopg2==2.9.9
black==24.4.0
envparse==0.2.0
python-jose==3.3.0
cryptography==42.0.5
bcrypt==4.1.2
email-validator==2.1.1
pre-commit==3.7.0
flake8==7.0.0
pytest==8.1.1
pytest-asyncio==0.21.1
httpx==0.27.0